add functionality to buttons below the blog article
