+++ nextjs/src/components/Blog/BlogPage.tsx
@@ -2,7 +2,6 @@ import Markdown from "react-markdown";
 import rehypeRaw from "rehype-raw";
 import remarkGfm from "remark-gfm";
 import rehypeSlug from "rehype-slug";
-import { ArrowUp } from "lucide-react";
 
 import { MarkdownComponents } from "../Markdown";
 import { DateToLocal } from "../DateToLocal";
@@ -11,8 +10,8 @@ import { Blog, Toc, User } from "@/types";
 // smooth scroll to anchor links
 import "../../styles.css";
 import { TableOfContentsMobile } from "./TableOfContentsMobile";
-import { ShareButton } from "../ShareButton";
-import { VscShare } from "react-icons/vsc";
+import { ShareButton } from "./ShareButton";
+import Buttons from "./Buttons";
 
 export const BlogPage = ({ blog, user, toc }: { blog: Blog; user: User | null; toc: Toc[] }) => {
     const blogUrl = `${process.env.NEXT_PUBLIC_SITE_URL}/blog/${blog.id}`;
@@ -54,16 +53,7 @@ export const BlogPage = ({ blog, user, toc }: { blog: Blog; user: User | null; t
                 >
                     {blog.content}
                 </Markdown>
-                <div className="mx-auto my-5 flex justify-end gap-5">
-                    <button className="flex hover:bg-blue-500 border border-black px-4 py-2 sm:px-10 sm:py-2 text-sm rounded-full bg-white dark:bg-black dark:text-white dark:border-white dark:hover:text-blue-500">
-                        <VscShare className="w-5 h-5" />
-                        <span className="ml-3">Share</span>
-                    </button>
-                    <button className="flex hover:bg-blue-500 border border-black px-4 py-2 sm:px-10 sm:py-2 text-sm rounded-full bg-white dark:bg-black dark:text-white dark:border-white dark:hover:text-blue-500">
-                        <ArrowUp className="w-5 h-5" />
-                        <span className="ml-3">Back to top</span>
-                    </button>
-                </div>
+                <Buttons title={blog.title} text={blog.description} url={blogUrl} />
             </div>
         </div>
     );
